+++ src/progress/methods.rs
@@ -26,3 +26,77 @@ pub trait BarMethods {
     /// Print a message via bar (without overlap with bars).
     fn write<T: Into<String>>(&mut self, text: T);
 }
+
+#[macro_export]
+#[doc(hidden)]
+macro_rules! _impl_bar_methods {
+    ($struct: ident, $render: ident) => {
+        impl $crate::BarMethods for $struct {
+            fn clear(&mut self) {
+                self.pb.clear();
+            }
+
+            fn input<T: Into<String>>(&mut self, text: T) -> Result<String, std::io::Error> {
+                self.clear();
+                self.pb.writer.print_str(&text.into());
+
+                let mut input_string = String::new();
+                std::io::stdin().read_line(&mut input_string)?;
+
+                if self.pb.leave {
+                    self.refresh();
+                }
+
+                Ok(input_string)
+            }
+
+            fn refresh(&mut self) {
+                if !self.pb.force_refresh {
+                    self.pb.force_refresh = true;
+                    self.update(0);
+                    self.pb.force_refresh = false;
+                } else {
+                    self.update(0);
+                }
+            }
+
+            fn render(&mut self) -> String {
+                $render(self)
+            }
+
+            fn reset(&mut self, total: Option<usize>) {
+                self.pb.reset(total);
+            }
+
+            fn update(&mut self, n: usize) {
+                self.pb.init();
+
+                if self.pb.trigger(n) {
+                    let text = self.render();
+                    let length = $crate::term::Colorizer::len_ansi(text.as_str()) as i16;
+
+                    if length != self.pb.bar_length {
+                        self.pb.clear();
+                    }
+
+                    self.pb.bar_length = length;
+                    self.pb.write_at(text);
+                }
+            }
+
+            fn update_to(&mut self, update_to_n: usize) {
+                self.pb.n = update_to_n;
+                self.update(0);
+            }
+
+            fn write<T: Into<String>>(&mut self, text: T) {
+                self.pb.clear();
+                self.pb.writer.print(format_args!("\r{}\n", text.into()));
+
+                if self.pb.leave {
+                    self.refresh();
+                }
+            }
+        }
+    };
+}