+++ src/progress/mod.rs
@@ -1,11 +1,9 @@
 mod bar;
 mod iterator;
-mod macros;
 mod methods;
 mod rich;
 
 pub use bar::*;
 pub use iterator::*;
-pub use macros::*;
 pub use methods::*;
 pub use rich::*;