+++ src/term/mod.rs
@@ -2,10 +2,8 @@
 
 mod colours;
 mod env;
-mod string;
 mod writer;
 
 pub use colours::*;
 pub use env::*;
-pub use string::*;
 pub use writer::*;